requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
channels==4.3.1
Pillow==11.3.0
